+++ src/chatautobuttons.ts
@@ -3,6 +3,7 @@ import { startDiveAndBreach } from "./spells/diveandbreach.ts";
 import { playRisingHurricaneAtLastPlacedTemplate } from "./actions/risinghurricane.ts";
 import { ChatMessagePF2e, TokenPF2e } from "foundry-pf2e";
 import { startSonicDash } from "./actions/sonicdash.ts";
+import { startTectonicStomp } from "./actions/tectonicstomp.ts";
 
 const SLUG_PREFIX = 'origin:item:slug:';
 
@@ -30,6 +31,10 @@ const AUTO_BUTTONS_ACTIONS: Record<string, ButtonSpec> = {
     "sonic-dash": {
         label: "Start dashing!",
         function: startSonicDash
+    },
+    "tectonic-stomp": {
+        label: "Start stomping!",
+        function: startTectonicStomp
     }
 }